Client TCP Sketch
I've left comments in the main() that outline the very basic things
that will be involved in the protocol. It doesn't take into account any
of the iPod specific functions. I'll try and come up with something
similar for the Server. We'll need to make some more specific decisions
about the protocol soon.